Create randomly_populate_display
This is the code to solve Micro::Bit Problem 2. A nested while loop with if-else statements turns up the brightness of the LED in location (x,y) on the Micro::Bit display, enabling the value to change from 0 (lowest LED setting - off) to 255 (brightest LED setting - on). After the nested while loop is finished executing and the incremented variable count reaches 25, the display resets all 25 of the (x,y) coordinates back to zero, and the process repeats itself again and again.