Create move_led
This is a slightly modified version of a classmate's solution to Problem 4, which is to randomly plot a LED on the Micro::Bit display and move the light using the A and B buttons. Button A moves LED one space to the left on the x-axis, Button B moves LED one space to the right on the x-axis, and Buttons A and B pushed together moves the LED up one space each click. When the LED reaches the outside of the display, the light jumps to the first space on the opposite side of where the LED was last lit up.