Create send_temp_radio
This code gets a temperature reading using the built-in temperature sensor on the board and uses radio to send the data to a second, separate board. When Button A is triggered on the Micro::Bit, the display scrolls the temperature in Celcius and when Button B is triggered, the temperature is given in Fahrenheit.